Refactor motor mapping into the shared lib
