+++ src/bin/sns-teleopd.c
@@ -50,88 +50,7 @@
 #include "sns.h"
 #include "sns/joystick/gamepad.h"
 #include "sns/event.h"
-
-struct sns_motor_map {
-    size_t n;
-    const struct aa_rx_sg *sg;
-    char **name;
-    aa_rx_config_id *id;
-};
-
-void
-sns_motor_map_in( const struct sns_motor_map *M,
-                  size_t n_sub, const double *q_sub,
-                  double *q_all )
-{
-    size_t n = AA_MIN(n_sub, M->n);
-    for( size_t i = 0; i < n; i ++ ) {
-        q_all[ M->id[i] ] = q_sub[i];
-    }
-}
-
-void
-sns_motor_map_destroy( struct sns_motor_map *m )
-{
-    for( size_t i = 0; i < m->n; i ++ ) {
-        free(m->name[i]);
-    }
-    free(m->name);
-    free(m->id);
-    free(m);
-}
-
-struct sns_motor_map *
-sns_motor_map_parse( const char *str )
-{
-    char *saveptr = NULL;
-    size_t n_var = 0;
-    size_t n_str = 0;
-
-    /* Count vars and string length */
-    if( '\0' != *str ) n_var++;
-    while( '\0' != str[n_str] ) {
-        if( ',' == str[n_str] ) n_var++;
-        n_str++;
-    }
-
-    char buf[1 + n_str];
-    strcpy(buf,str);
-
-    struct sns_motor_map *M = AA_NEW(struct sns_motor_map);
-    M->n = n_var;
-    M->name = AA_NEW_AR(char*,M->n);
-    M->id = AA_NEW_AR(aa_rx_config_id,M->n);
-
-    /* tokenize string */
-    {
-        size_t i = 0;
-        char *tok = strtok_r(buf, ",", &saveptr);
-        while(tok) {
-            assert( i < M->n );
-            M->name[i] = strdup(tok);
-            M->id[i] = AA_RX_FRAME_NONE;
-            i++;
-            tok = strtok_r(NULL, ",", &saveptr);
-        }
-    }
-
-    return M;
-
-}
-
-int
-sns_motor_map_fill_id( const struct aa_rx_sg *sg, struct sns_motor_map *M )
-{
-    M->sg = sg;
-    for( size_t i = 0; i < M->n; i ++ ) {
-        M->id[i] = aa_rx_sg_config_id(sg,M->name[i]);
-        if( M->id[i] < 0 ) {
-            SNS_LOG( LOG_ERR, "Unknown configuration variable: `%s'", M->name[i] );
-            return  -1;
-        }
-    }
-    return 0;
-}
+#include "sns/motor.h"
 
 
 struct joint_ctrl {
